requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Werkzeug==2.3.7
requests==2.31.0
oauthlib==3.2.2
WeasyPrint==60.1
python-dotenv==1.0.0
gunicorn==21.2.0
email-validator==2.1.0.post1
PyJWT==2.8.0
SQLAlchemy==2.0.23
Flask-Limiter==3.5.0
limits==3.6.0
pandas==2.1.3
